refactor(tasks): :truck: move 'check_workflow' task
